Fix bugs, clean code and implement color
I changed a few details in polygon_soup_item, like the sizeof(float) or the size of the location array.

Fixed the bug that deformed an item in polyhedron_item, caused by glGet(GL_POLYGON_MODE) that takes an array of 2 int and not a single int.

Implemented the color changes in polyhedron_item depending of the selection.